Backend/requirements.txt
# Core
Django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-allauth[socialaccount]>=0.61
python-dotenv>=1.0

# Image / colour
Pillow>=10.0
colorthief>=0.2.1

# Hugging Face
huggingface_hub>=0.22
requests>=2.31

# Anthropic
anthropic>=0.19.0

# Image processing (anchor feature)
rembg>=2.0
opencv-python-headless>=4.9

# Dev / prod DB
dj-database-url>=2.1.0
psycopg2-binary>=2.9
